Extract BeachSkyline into its own file
